requirements.txt
discord.py
python-dotenv
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-mock==3.12.0
discord.py==2.3.2
python-dotenv==1.0.0